+++ src/controllers/WebhooksController.php
@@ -8,8 +8,13 @@
 
 namespace workingconcept\snipcart\controllers;
 
+use workingconcept\snipcart\events\CustomerEvent;
+use workingconcept\snipcart\events\OrderStatusEvent;
+use workingconcept\snipcart\events\OrderTrackingEvent;
+use workingconcept\snipcart\events\SubscriptionEvent;
+use workingconcept\snipcart\models\Subscription;
 use workingconcept\snipcart\Snipcart;
-use workingconcept\snipcart\events\WebhookEvent;
+use workingconcept\snipcart\events\OrderEvent;
 use workingconcept\snipcart\records\WebhookLog;
 use workingconcept\snipcart\records\ShippingQuoteLog;
 use workingconcept\snipcart\models\Order;
@@ -26,16 +31,67 @@ class WebhooksController extends Controller
     // =========================================================================
 
     /**
-     * @event WebhookEvent Triggered before a completed event is handled by the plugin.
+     * @event OrderEvent Triggered before a completed event is handled by the plugin.
      */
     const EVENT_BEFORE_PROCESS_COMPLETED_ORDER = 'beforeProcessCompletedOrder';
-    
+
+    /**
+     * @event OrderStatusEvent
+     */
+    const EVENT_ON_ORDER_STATUS_CHANGED = 'onOrderStatusChanged';
+
+    /**
+     * @event OrderStatusEvent
+     */
+    const EVENT_ON_ORDER_PAYMENT_STATUS_CHANGED = 'onOrderPaymentStatusChanged';
+
+    /**
+     * @event OrderTrackingEvent
+     */
+    const EVENT_ON_ORDER_TRACKING_CHANGED = 'onOrderTrackingChanged';
+
+    /**
+     * @event SubscriptionEvent
+     */
+    const EVENT_ON_SUBSCRIPTION_CREATED = 'onSubscriptionCreated';
+
+    /**
+     * @event SubscriptionEvent
+     */
+    const EVENT_ON_SUBSCRIPTION_CANCELLED = 'onSubscriptionCancelled';
+
+    /**
+     * @event SubscriptionEvent
+     */
+    const EVENT_ON_SUBSCRIPTION_PAUSED = 'onSubscriptionPaused';
+
+    /**
+     * @event SubscriptionEvent
+     */
+    const EVENT_ON_SUBSCRIPTION_RESUMED = 'onSubscriptionResumed';
+
+    /**
+     * @event SubscriptionEvent
+     */
+    const EVENT_ON_SUBSCRIPTION_INVOICE_CREATED = 'onSubscriptionInvoiceCreated';
+
+    /**
+     * @event TaxesEvent
+     */
+    const EVENT_ON_TAXES_CALCULATE = 'onTaxesCalculate';
+
+    /**
+     * @event CustomerEvent
+     */
+    const EVENT_ON_CUSTOMER_UPDATE = 'onCustomerUpdate';
+
     /**
      * Snipcart's available webhook events
      */
     const WEBHOOK_ORDER_COMPLETED               = 'order.completed';
     const WEBHOOK_SHIPPINGRATES_FETCH           = 'shippingrates.fetch';
     const WEBHOOK_ORDER_STATUS_CHANGED          = 'order.status.changed';
+    const WEBHOOK_ORDER_PAYMENT_STATUS_CHANGED  = 'order.paymentStatus.changed';
     const WEBHOOK_ORDER_TRACKING_NUMBER_CHANGED = 'order.trackingNumber.changed';
     const WEBHOOK_SUBSCRIPTION_CREATED          = 'subscription.created';
     const WEBHOOK_SUBSCRIPTION_CANCELLED        = 'subscription.cancelled';
@@ -48,15 +104,16 @@ class WebhooksController extends Controller
     const WEBHOOK_EVENT_MAP = [
         self::WEBHOOK_ORDER_COMPLETED               => '_handleOrderCompleted',
         self::WEBHOOK_SHIPPINGRATES_FETCH           => '_handleShippingRatesFetch',
-        self::WEBHOOK_ORDER_STATUS_CHANGED          => '_notSupportedResponse',
-        self::WEBHOOK_ORDER_TRACKING_NUMBER_CHANGED => '_notSupportedResponse',
-        self::WEBHOOK_SUBSCRIPTION_CREATED          => '_notSupportedResponse',
-        self::WEBHOOK_SUBSCRIPTION_CANCELLED        => '_notSupportedResponse',
-        self::WEBHOOK_SUBSCRIPTION_PAUSED           => '_notSupportedResponse',
-        self::WEBHOOK_SUBSCRIPTION_RESUMED          => '_notSupportedResponse',
-        self::WEBHOOK_SUBSCRIPTION_INVOICE_CREATED  => '_notSupportedResponse',
-        self::WEBHOOK_TAXES_CALCULATE               => '_notSupportedResponse',
-        self::WEBHOOK_CUSTOMER_UPDATED              => '_notSupportedResponse',
+        self::WEBHOOK_ORDER_STATUS_CHANGED          => '_handleOrderStatusChange',
+        self::WEBHOOK_ORDER_PAYMENT_STATUS_CHANGED  => '_handleOrderPaymentStatusChange',
+        self::WEBHOOK_ORDER_TRACKING_NUMBER_CHANGED => '_handleOrderTrackingNumberChange',
+        self::WEBHOOK_SUBSCRIPTION_CREATED          => '_handleSubscriptionCreated',
+        self::WEBHOOK_SUBSCRIPTION_CANCELLED        => '_handleSubscriptionCancelled',
+        self::WEBHOOK_SUBSCRIPTION_PAUSED           => '_handleSubscriptionPaused',
+        self::WEBHOOK_SUBSCRIPTION_RESUMED          => '_handleSubscriptionResumed',
+        self::WEBHOOK_SUBSCRIPTION_INVOICE_CREATED  => '_handleOrderSubscriptionInvoiceCreated',
+        self::WEBHOOK_TAXES_CALCULATE               => '_handleTaxesCalculate',
+        self::WEBHOOK_CUSTOMER_UPDATED              => '_handleCustomerUpdated',
     ];
 
     const WEBHOOK_MODE_LIVE = 'Live';
@@ -70,7 +127,7 @@ class WebhooksController extends Controller
      * @inheritdoc
      * @var bool Disable CSRF for this controller
      */
-    public $enableCsrfValidation = false; //
+    public $enableCsrfValidation = false;
 
     /**
      * @inheritdoc
@@ -82,6 +139,11 @@ class WebhooksController extends Controller
      * @var bool
      */
     private static $validateWebhook = true;
+
+    /**
+     * @var mixed local reference to decoded post data
+     */
+    private $postData;
 
 
     // Public Methods
@@ -116,19 +178,19 @@ class WebhooksController extends Controller
          */
         $this->requirePostRequest();
 
-        $postData = json_decode(Craft::$app->getRequest()->getRawBody());
+        $this->postData = json_decode(Craft::$app->getRequest()->getRawBody());
 
-        if ($reason = $this->_hasInvalidRequestData($postData))
+        if ($reason = $this->_hasInvalidRequestData())
         {
             return $this->_badRequestResponse([ 'reason' => $reason ]);
         }
 
         if (Snipcart::$plugin->getSettings()->logWebhookRequests)
         {
-            $this->_logWebhookTransaction($postData);
+            $this->_logWebhookTransaction();
         }
 
-        return $this->_handleWebhookData($postData->eventName, $postData->content);
+        return $this->_handleWebhookData($this->postData->eventName);
     }
 
 
@@ -138,11 +200,10 @@ class WebhooksController extends Controller
     /**
      * Send the webhook's post content to the appropriate handler method.
      * @param string $eventName    Event name from `WEBHOOK_EVENT_MAP`.
-     * @param mixed  $postContent  Decoded post data ->content.
      * @return Response
      * @throws Exception if the mapped handler method doesn't exist.
      */
-    private function _handleWebhookData($eventName, $postContent): Response
+    private function _handleWebhookData($eventName): Response
     {
         if (method_exists($this, self::WEBHOOK_EVENT_MAP[$eventName]))
         {
@@ -150,7 +211,7 @@ class WebhooksController extends Controller
              * Call the method defined in `WEBHOOK_EVENT_MAP`,
              * with $postContent as its argument.
              */
-            return $this->{self::WEBHOOK_EVENT_MAP[$eventName]}($postContent);
+            return $this->{self::WEBHOOK_EVENT_MAP[$eventName]}();
         }
 
         throw new Exception(sprintf(
@@ -160,15 +221,13 @@ class WebhooksController extends Controller
     }
 
     /**
-     * Handle the completed order.
-     *
-     * @param mixed $postContent Decoded post data ->content.
+     * Handle a completed order.
      * @return Response
      * @throws
      */
-    private function _handleOrderCompleted($postContent): Response
+    private function _handleOrderCompleted(): Response
     {
-        $order = new Order($postContent);
+        $order = new Order($this->postData->content);
 
         $responseData = [
             'success' => true,
@@ -179,7 +238,7 @@ class WebhooksController extends Controller
         {
             $this->trigger(
                 self::EVENT_BEFORE_PROCESS_COMPLETED_ORDER,
-                new WebhookEvent([
+                new OrderEvent([
                     'order' => $order
                 ])
             );
@@ -229,13 +288,11 @@ class WebhooksController extends Controller
     /**
      * Process Snipcart's shipping rate event, which gives us order details
      * and lets us send back shipping options.
-     *
-     * @param mixed $postContent Decoded post data ->content.
      * @return Response
      */
-    private function _handleShippingRatesFetch($postContent): Response
+    private function _handleShippingRatesFetch(): Response
     {
-        $order    = new Order($postContent);
+        $order    = new Order($this->postData->content);
         $rates    = Snipcart::$plugin->shipments->collectRatesForOrder($order);
         $response = $this->asJson($rates);
 
@@ -249,6 +306,224 @@ class WebhooksController extends Controller
         }
 
         return $response;
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleOrderStatusChange(): Response
+    {
+        $fromStatus = $this->postData->from;
+        $toStatus   = $this->postData->to;
+        $order      = new Order($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_ORDER_STATUS_CHANGED))
+        {
+            $this->trigger(
+                self::EVENT_ON_ORDER_STATUS_CHANGED,
+                new OrderStatusEvent([
+                    'order'      => $order,
+                    'fromStatus' => $fromStatus,
+                    'toStatus'   => $toStatus,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleOrderPaymentStatusChange(): Response
+    {
+        $fromStatus = $this->postData->from;
+        $toStatus   = $this->postData->to;
+        $order      = new Order($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_ORDER_PAYMENT_STATUS_CHANGED))
+        {
+            $this->trigger(
+                self::EVENT_ON_ORDER_PAYMENT_STATUS_CHANGED,
+                new OrderStatusEvent([
+                    'order'      => $order,
+                    'fromStatus' => $fromStatus,
+                    'toStatus'   => $toStatus,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleOrderTrackingNumberChange(): Response
+    {
+        $trackingNumber = $this->postData->trackingNumber;
+        $trackingUrl    = $this->postData->trackingUrl;
+        $order          = new Order($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_ORDER_TRACKING_CHANGED))
+        {
+            $this->trigger(
+                self::EVENT_ON_ORDER_TRACKING_CHANGED,
+                new OrderTrackingEvent([
+                    'order'          => $order,
+                    'trackingNumber' => $trackingNumber,
+                    'trackingUrl'    => $trackingUrl,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleSubscriptionCreated(): Response
+    {
+        $subscription = new Subscription($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_SUBSCRIPTION_CREATED))
+        {
+            $this->trigger(
+                self::EVENT_ON_SUBSCRIPTION_CREATED,
+                new SubscriptionEvent([
+                    'subscription' => $subscription,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleSubscriptionCancelled(): Response
+    {
+        $subscription = new Subscription($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_SUBSCRIPTION_CANCELLED))
+        {
+            $this->trigger(
+                self::EVENT_ON_SUBSCRIPTION_CANCELLED,
+                new SubscriptionEvent([
+                    'subscription' => $subscription,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleSubscriptionPaused(): Response
+    {
+        $subscription = new Subscription($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_SUBSCRIPTION_PAUSED))
+        {
+            $this->trigger(
+                self::EVENT_ON_SUBSCRIPTION_PAUSED,
+                new SubscriptionEvent([
+                    'subscription' => $subscription,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleSubscriptionResumed(): Response
+    {
+        $subscription = new Subscription($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_SUBSCRIPTION_RESUMED))
+        {
+            $this->trigger(
+                self::EVENT_ON_SUBSCRIPTION_RESUMED,
+                new SubscriptionEvent([
+                    'subscription' => $subscription,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleSubscriptionInvoiceCreated(): Response
+    {
+        $subscription = new Subscription($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_SUBSCRIPTION_INVOICE_CREATED))
+        {
+            $this->trigger(
+                self::EVENT_ON_SUBSCRIPTION_INVOICE_CREATED,
+                new SubscriptionEvent([
+                    'subscription' => $subscription,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleTaxesCalculate(): Response
+    {
+        $order = new Order($this->postData->content);
+        $taxes = [];
+
+        if ($this->hasEventHandlers(self::EVENT_ON_TAXES_CALCULATE))
+        {
+            $this->trigger(
+                self::EVENT_ON_TAXES_CALCULATE,
+                new SubscriptionEvent([
+                    'order' => $order,
+                    'taxes' => [],
+                ])
+            );
+
+            $taxes = array_merge($taxes, $event->taxes);
+        }
+
+        return $this->asJson([
+            'taxes' => $taxes
+        ]);
+    }
+
+    /**
+     * @return Response
+     */
+    private function _handleCustomerUpdated(): Response
+    {
+        $customer = new Customer($this->postData->content);
+
+        if ($this->hasEventHandlers(self::EVENT_ON_CUSTOMER_UPDATE))
+        {
+            $this->trigger(
+                self::EVENT_ON_CUSTOMER_UPDATE,
+                new CustomerEvent([
+                    'customer' => $customer,
+                ])
+            );
+        }
+
+        return $this->_notSupportedResponse();
     }
 
     /**
@@ -276,36 +551,33 @@ class WebhooksController extends Controller
     /**
      * Output a 200 response so Snipcart knows we're okay but not
      * handling the event.
-     * @param mixed $postContent Decoded post data ->content.
      * @return Response
      */
-    private function _notSupportedResponse($postContent): Response
+    private function _notSupportedResponse(): Response
     {
         return $this->asJson([ 'success' => true ]);
     }
 
     /**
      * Store webhook details to the database for later scrutiny.
-     * @param $postBody
      */
-    private function _logWebhookTransaction($postBody)
+    private function _logWebhookTransaction()
     {
         $webhookLog = new WebhookLog();
 
         $webhookLog->siteId = Craft::$app->sites->currentSite->id;
-        $webhookLog->type   = $postBody->eventName;
-        $webhookLog->body   = $postBody;
+        $webhookLog->type   = $this->postData->eventName;
+        $webhookLog->body   = $this->postData;
 
         $webhookLog->save();
     }
 
     /**
      * Make sure we don't have invalid data, or return a reason if we do.
-     * @param $postData
      * @return bool|string
      * @throws BadRequestHttpException if there's a problem with the token.
      */
-    private function _hasInvalidRequestData($postData)
+    private function _hasInvalidRequestData()
     {
         /**
          * Reject requests that can't be validated.
@@ -319,7 +591,7 @@ class WebhooksController extends Controller
          * Every Snipcart post should have an eventName, so we've got
          * missing data or a bad format.
          */
-        if ($postData === null || !isset($postData->eventName))
+        if ($this->postData === null || !isset($this->postData->eventName))
         {
             return 'NULL request body or missing eventName.';
         }
@@ -327,7 +599,7 @@ class WebhooksController extends Controller
         /**
          * Every Snipcart post should have a content property.
          */
-        if (!isset($postData->content))
+        if (!isset($this->postData->content))
         {
             return 'Request missing content.';
         }
@@ -336,7 +608,7 @@ class WebhooksController extends Controller
          * We've received an invalid `eventName`.
          */
         if (! array_key_exists(
-            $postData->eventName,
+            $this->postData->eventName,
             self::WEBHOOK_EVENT_MAP
         ))
         {