+++ src/events/OrderEvent.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Snipcart plugin for Craft CMS 3.x
+ *
+ * @link      https://workingconcept.com
+ * @copyright Copyright (c) 2018 Working Concept Inc.
+ */
+
+namespace workingconcept\snipcart\events;
+
+use workingconcept\snipcart\models\Order;
+use yii\base\Event;
+
+/**
+ * Order event class.
+ *
+ * @link      https://workingconcept.com
+ * @copyright Copyright (c) 2018 Working Concept Inc.
+ */
+class OrderEvent extends Event
+{
+    // Properties
+    // =========================================================================
+
+    /**
+     * @var Order
+     */
+    public $order;
+
+}